hooks: split hooks api to Info and Callbacks
This would allow sidecar to expose multiple versions of hook
implementation api on one socket.
