+++ pkg/hooks/info/api.proto
@@ -1,0 +1,25 @@
+syntax = "proto3";
+
+package kubevirt.hooks.info;
+
+service Info {
+  rpc Info (InfoParams) returns (InfoResult);
+}
+
+message InfoParams {}
+
+message InfoResult {
+  // name of the hook used by virt-launcher to compare it with requested hooks
+  string name = 1;
+  // hookPoints is a list of subscribed hook points
+  repeated HookPoint hookPoints = 3;
+  // versions is a list of implemented hook Callbacks service versions
+  repeated string versions = 4;
+}
+
+message HookPoint {
+  // name represents name of the subscribed hook point
+  string name = 1;
+  // priority is used to sort hooks prior to their execution (second key is the name)
+  uint32 priority = 2;
+}
