+++ pkg/hooks/info/info.go
@@ -1,0 +1,3 @@
+package info
+
+const OnDefineDomainHookPointName = "OnDefineDomain"
