--- pkg/hooks/v1alpha/api.proto
@@ -1,36 +1,0 @@
-syntax = "proto3";
-
-package kubevirt.hooks.v1alpha;
-
-service Hook {
-  rpc Info (InfoParams) returns (InfoResult);
-  rpc OnDefineDomain (OnDefineDomainParams) returns (OnDefineDomainResult);
-}
-
-message InfoParams {
-  // supportedVersions is a list of versions supported by client
-  repeated string supportedVersions = 1;
-}
-
-message InfoResult {
-  // name of the hook used by virt-launcher to compare it with requested hooks
-  string name = 1;
-  // priority is used to sort hooks prior to their execution (second key is the name)
-  uint32 priority = 2;
-  // hookPoints is a list of subscribed hook points
-  repeated string hookPoints = 3;
-  // version is choosen from passed supportedVersions, this version will be used for communication
-  string version = 4;
-}
-
-message OnDefineDomainParams {
-  // domainXML is original libvirt domain specification, it is XML passed as string
-  string domainXML = 1;
-  // vm is VirtualMachine is object of virtual machine currently processed by virt-launcher, it is encoded as JSON and passed as string
-  string vm = 2;
-}
-
-message OnDefineDomainResult {
-  // domainXML is processed libvirt domain specification, it is XML passed as string
-  string domainXML = 1;
-}
